Lots of changes
Implemented the privacy modal for the appointment.php. Click the link on the checkbox to access the modal.
In request.php, implemented an alert for approving an appointment time set by the seller through a button. Also for requests by buyer the seller can press a button to schedule an appointment time and day (links to appointment.php).
Started the about.php and privacty.php but the about page div won't turn into a scroll.

Be cautious! There were changes to the database's css files that I didn't do.  I'm not sure what happened. Please check.